# backend/indexer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/scan-tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of recent scan tasks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List recent scan tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by site ID",
                        "name": "site_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by domain (partial match)",
                        "name": "domain",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, processing, completed, failed, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ListTasksResponse"
                        }
                    }
                }
            }
        },
        "/api/scan-tasks/cancel": {
            "post": {
                "description": "Cancel pending or processing scan tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Cancel scan tasks",
                "parameters": [
                    {
                        "description": "Task IDs to cancel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CancelTasksRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CancelTasksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/scan-tasks/{id}": {
            "get": {
                "description": "Get scan task details and status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get scan task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.ScanTask"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/sites/{id}/all-urls": {
            "get": {
                "description": "Returns all URL strings for a site (for crawler deduplication)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "Get all discovered URLs for a site",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.AllURLsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/sites/{id}/pending-urls": {
            "get": {
                "description": "Get URLs that need to be parsed (status=pending, respects retry logic)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "Get pending URLs for crawling",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Max URLs to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.PendingURLsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/sites/{id}/sitemap-urls": {
            "get": {
                "description": "Get paginated list of URLs from sitemap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "List sitemap URLs for a site",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, indexed, error, skipped)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.SitemapURLsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sites/{id}/sitemap-urls/stats": {
            "get": {
                "description": "Get count of URLs by status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "Get sitemap URLs statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.SitemapURLStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_video-analitics_backend_pkg_status.Stage": {
            "description": "Task stage",
            "type": "string",
            "enum": [
                "sitemap",
                "page",
                "done"
            ],
            "x-enum-comments": {
                "StageDone": "all stages completed",
                "StagePage": "parsing pages",
                "StageSitemap": "collecting URLs from sitemap"
            },
            "x-enum-descriptions": [
                "collecting URLs from sitemap",
                "parsing pages",
                "all stages completed"
            ],
            "x-enum-varnames": [
                "StageSitemap",
                "StagePage",
                "StageDone"
            ]
        },
        "github_com_video-analitics_backend_pkg_status.Task": {
            "description": "Task status",
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-comments": {
                "TaskCancelled": "cancelled by user",
                "TaskCompleted": "finished successfully",
                "TaskFailed": "finished with error",
                "TaskPending": "waiting in queue",
                "TaskProcessing": "being processed by worker"
            },
            "x-enum-descriptions": [
                "waiting in queue",
                "being processed by worker",
                "finished successfully",
                "finished with error",
                "cancelled by user"
            ],
            "x-enum-varnames": [
                "TaskPending",
                "TaskProcessing",
                "TaskCompleted",
                "TaskFailed",
                "TaskCancelled"
            ]
        },
        "github_com_video-analitics_backend_pkg_status.URL": {
            "description": "URL indexing status",
            "type": "string",
            "enum": [
                "pending",
                "indexed",
                "error",
                "skipped"
            ],
            "x-enum-comments": {
                "URLError": "failed after max retries",
                "URLIndexed": "successfully indexed",
                "URLPending": "waiting to be indexed",
                "URLSkipped": "skipped (e.g., XML sitemap reference)"
            },
            "x-enum-descriptions": [
                "waiting to be indexed",
                "successfully indexed",
                "failed after max retries",
                "skipped (e.g., XML sitemap reference)"
            ],
            "x-enum-varnames": [
                "URLPending",
                "URLIndexed",
                "URLError",
                "URLSkipped"
            ]
        },
...
        "github_com_video-analitics_indexer_internal_repo.ScanTask": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "page_result": {
                    "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.StageResult"
                },
                "site_id": {
                    "type": "string"
                },
                "sitemap_result": {
                    "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.StageResult"
                },
                "stage": {
                    "$ref": "#/definitions/github_com_video-analitics_backend_pkg_status.Stage"
                },
                "status": {
                    "$ref": "#/definitions/github_com_video-analitics_backend_pkg_status.Task"
                }
            }
        },
...
        "github_com_video-analitics_indexer_internal_repo.SitemapURL": {
            "type": "object",
            "properties": {
                "changefreq": {
                    "type": "string"
                },
                "depth": {
                    "description": "Глубина: 0 = из sitemap/главная, 1-3 = найдены при парсинге страниц",
                    "type": "integer"
                },
                "discovered_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "indexed_at": {
                    "type": "string"
                },
                "is_xml": {
                    "type": "boolean"
                },
                "last_attempt_at": {
                    "type": "string"
                },
                "lastmod": {
                    "type": "string"
                },
                "locked_until": {
                    "type": "string"
                },
                "priority": {
                    "type": "number"
                },
                "retry_count": {
                    "type": "integer"
                },
                "site_id": {
                    "type": "string"
                },
                "sitemap_source": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_video-analitics_backend_pkg_status.URL"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_video-analitics_indexer_internal_repo.SitemapURLStats": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "indexed": {
                    "type": "integer"
                },
                "pending": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_video-analitics_indexer_internal_repo.StageResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "failed": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_video-analitics_backend_pkg_status.Task"
                },
                "success": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "indexer_internal_handler.AllURLsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "indexer_internal_handler.CancelTasksRequest": {
            "type": "object",
            "properties": {
                "task_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "indexer_internal_handler.CancelTasksResponse": {
            "type": "object",
            "properties": {
                "cancelled_count": {
                    "type": "integer"
                }
            }
        },
...
        "indexer_internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "indexer_internal_handler.ListTasksResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.ScanTask"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "indexer_internal_handler.PendingURLWithDepth": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "indexer_internal_handler.PendingURLsResponse": {
            "type": "object",
            "properties": {
                "all_indexed": {
                    "type": "boolean"
                },
                "in_retry": {
                    "type": "boolean"
                },
                "indexed_urls": {
                    "type": "integer"
                },
                "total_urls": {
                    "type": "integer"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/indexer_internal_handler.PendingURLWithDepth"
                    }
                }
            }
        },
...
        "indexer_internal_handler.SitemapURLsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_video-analitics_indexer_internal_repo.SitemapURL"
                    }
                }
            }
        },
...
    }
}